[MODEL-8962] Modularize adapters
