+++ custom_model_runner/datarobot_drum/drum/adapters/cli/shared/drum_class_label_adapter.py
@@ -6,11 +6,15 @@ Released under the terms of DataRobot Tool and Utility Agreement.
 """
 import sys
 
+import logging
 import numpy as np
 import pandas as pd
 
-from datarobot_drum.drum.enum import TargetType
 from datarobot_drum.drum.exceptions import DrumCommonException
+from datarobot_drum.drum.enum import LOGGER_NAME_PREFIX, TargetType
+
+
+logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
 
 
 def needs_class_labels(
@@ -86,3 +90,61 @@ def infer_class_labels(target_type, input_filename, target_filename=None, target
             )
 
     return possible_class_labels
+
+
+class DrumClassLabelAdapter(object):
+    """
+    Shared parent class for other adapters that handles parsing class label input parameters.
+    """
+
+    def __init__(
+        self, target_type, positive_class_label=None, negative_class_label=None, class_labels=None,
+    ):
+        """
+        Parameters
+        ----------
+        target_type: datarobot_drum.drum.enum.TargetType
+        positive_class_label: str or None
+            Optional. Name of the positive class label if target type is binary
+        negative_class_label: str or None
+            Optional. Name of the negative class label if target type is binary
+        class_labels: list[str] or None
+            Optional. List of class labels
+        """
+        self.target_type = target_type
+        self.positive_class_label = positive_class_label
+        self.negative_class_label = negative_class_label
+        self.class_labels = class_labels
+
+    @property
+    def class_ordering(self):
+        if self.negative_class_label is not None and self.positive_class_label is not None:
+            class_order = [self.positive_class_label, self.negative_class_label]
+        elif self.class_labels:
+            class_order = self.class_labels
+        else:
+            class_order = None
+
+        return class_order
+
+    def _infer_class_labels_if_not_provided(
+        self, input_filename, target_filename=None, target_name=None
+    ):
+        if needs_class_labels(
+            target_type=self.target_type,
+            negative_class_label=self.negative_class_label,
+            positive_class_label=self.positive_class_label,
+            class_labels=self.class_labels,
+        ):
+            # TODO: Only pass y in here as a dataframe
+            class_labels = infer_class_labels(
+                target_type=self.target_type,
+                input_filename=input_filename,
+                target_filename=target_filename,
+                target_name=target_name,
+            )
+
+            if self.target_type == TargetType.BINARY:
+                self.positive_class_label, self.negative_class_label = class_labels
+            elif self.target_type == TargetType.MULTICLASS:
+                self.class_labels = class_labels
