+++ custom_model_runner/datarobot_drum/resource/components/Python/generic_predictor/generic_predictor.py
@@ -9,9 +9,8 @@ import urllib
 from typing import Optional
 
 import werkzeug
-from pandas import DataFrame
 
-from datarobot_drum.drum.adapters.drum_cli_adapter import DrumCLIAdapter
+from datarobot_drum.drum.adapters.cli.drum_score_adapter import DrumScoreAdapter
 from datarobot_drum.drum.enum import (
     LOGGER_NAME_PREFIX,
     TARGET_TYPE_ARG_KEYWORD,
@@ -25,8 +24,6 @@ from datarobot_drum.resource.unstructured_helpers import (
     _resolve_outgoing_unstructured_data,
 )
 
-from datarobot_drum.drum.utils.structured_input_read_utils import StructuredInputReadUtils
-
 from mlpiper.components.connectable_component import ConnectableComponent
 
 
@@ -36,14 +33,14 @@ class GenericPredictorComponent(ConnectableComponent):
         self.logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
         self._run_language = None
         self._predictor = None
-        self.cli_adapter: Optional[DrumCLIAdapter] = None
+        self.cli_adapter: Optional[DrumScoreAdapter] = None
 
     def configure(self, params):
         super(GenericPredictorComponent, self).configure(params)
         self._run_language = RunLanguage(params.get("run_language"))
 
         # Input filename is available at configuration time, so include it in the CLI adapter here.
-        self.cli_adapter = DrumCLIAdapter(
+        self.cli_adapter = DrumScoreAdapter(
             custom_task_folder_path=params["__custom_model_path__"],
             input_filename=params["input_filename"],
             sparse_column_filename=params.get("sparse_column_file"),
